Add column reodering in Customize Columns modal
